Retire usage of the term "interceptor".
"Interceptor" is a Resteasy 2 implementation specific term.  The
standard term as of JAX RS 2.0 is "filter".
