+++ common/src/main/java/org/candlepin/common/resteasy/filter/LinkHeaderResponseFilter.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.common.resteasy.interceptor;
+package org.candlepin.common.resteasy.filter;
 
 import org.candlepin.common.config.Configuration;
 import org.candlepin.common.paging.Page;
