+++ gutterball/src/main/java/org/candlepin/gutterball/guice/GutterballModule.java
@@ -36,9 +36,9 @@ import org.candlepin.common.exceptions.mappers.ValidationExceptionMapper;
 import org.candlepin.common.exceptions.mappers.WebApplicationExceptionMapper;
 import org.candlepin.common.exceptions.mappers.WriterExceptionMapper;
 import org.candlepin.common.resteasy.auth.NoOpFilter;
-import org.candlepin.common.resteasy.interceptor.DynamicJsonFilter;
-import org.candlepin.common.resteasy.interceptor.LinkHeaderResponseFilter;
-import org.candlepin.common.resteasy.interceptor.PageRequestFilter;
+import org.candlepin.common.resteasy.filter.DynamicJsonFilter;
+import org.candlepin.common.resteasy.filter.LinkHeaderResponseFilter;
+import org.candlepin.common.resteasy.filter.PageRequestFilter;
 import org.candlepin.common.validation.CandlepinMessageInterpolator;
 import org.candlepin.gutterball.config.ConfigProperties;
 import org.candlepin.gutterball.curator.ComplianceSnapshotCurator;
@@ -56,8 +56,8 @@ import org.candlepin.gutterball.report.StatusTrendReport;
 import org.candlepin.gutterball.resource.ReportsResource;
 import org.candlepin.gutterball.resource.StatusResource;
 import org.candlepin.gutterball.resteasy.JsonProvider;
-import org.candlepin.gutterball.resteasy.interceptor.OAuthFilter;
-import org.candlepin.gutterball.resteasy.interceptor.SecurityHoleFeature;
+import org.candlepin.gutterball.resteasy.filter.OAuthFilter;
+import org.candlepin.gutterball.resteasy.filter.SecurityHoleFeature;
 import org.candlepin.gutterball.util.EventHandlerLoader;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
