+++ gutterball/src/main/java/org/candlepin/gutterball/resteasy/filter/OAuthFilter.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.gutterball.resteasy.interceptor;
+package org.candlepin.gutterball.resteasy.filter;
 
 import org.candlepin.common.config.Configuration;
 import org.candlepin.common.exceptions.BadRequestException;
