+++ server/src/main/java/org/candlepin/guice/CandlepinModule.java
@@ -41,9 +41,9 @@ import org.candlepin.common.exceptions.mappers.ValidationExceptionMapper;
 import org.candlepin.common.exceptions.mappers.WebApplicationExceptionMapper;
 import org.candlepin.common.exceptions.mappers.WriterExceptionMapper;
 import org.candlepin.common.guice.JPAInitializer;
-import org.candlepin.common.resteasy.interceptor.DynamicJsonFilter;
-import org.candlepin.common.resteasy.interceptor.LinkHeaderResponseFilter;
-import org.candlepin.common.resteasy.interceptor.PageRequestFilter;
+import org.candlepin.common.resteasy.filter.DynamicJsonFilter;
+import org.candlepin.common.resteasy.filter.LinkHeaderResponseFilter;
+import org.candlepin.common.resteasy.filter.PageRequestFilter;
 import org.candlepin.common.validation.CandlepinMessageInterpolator;
 import org.candlepin.config.ConfigProperties;
 import org.candlepin.controller.CandlepinPoolManager;
@@ -106,14 +106,14 @@ import org.candlepin.resource.SubscriptionResource;
 import org.candlepin.resource.UserResource;
 import org.candlepin.resteasy.JsonProvider;
 import org.candlepin.resteasy.ResourceLocatorMap;
-import org.candlepin.resteasy.interceptor.AuthenticationFilter;
-import org.candlepin.resteasy.interceptor.AuthorizationFeature;
-import org.candlepin.resteasy.interceptor.PinsetterAsyncInterceptor;
-import org.candlepin.resteasy.interceptor.SecurityHoleAuthorizationFilter;
-import org.candlepin.resteasy.interceptor.StoreFactory;
-import org.candlepin.resteasy.interceptor.SuperAdminAuthorizationFilter;
-import org.candlepin.resteasy.interceptor.VerifyAuthorizationFilter;
-import org.candlepin.resteasy.interceptor.VersionResponseFilter;
+import org.candlepin.resteasy.filter.AuthenticationFilter;
+import org.candlepin.resteasy.filter.AuthorizationFeature;
+import org.candlepin.resteasy.filter.PinsetterAsyncFilter;
+import org.candlepin.resteasy.filter.SecurityHoleAuthorizationFilter;
+import org.candlepin.resteasy.filter.StoreFactory;
+import org.candlepin.resteasy.filter.SuperAdminAuthorizationFilter;
+import org.candlepin.resteasy.filter.VerifyAuthorizationFilter;
+import org.candlepin.resteasy.filter.VersionResponseFilter;
 import org.candlepin.service.UniqueIdGenerator;
 import org.candlepin.service.impl.DefaultUniqueIdGenerator;
 import org.candlepin.sync.ConsumerExporter;
@@ -302,7 +302,7 @@ public class CandlepinModule extends AbstractModule {
 
     private void configureInterceptors() {
         bind(PageRequestFilter.class);
-        bind(PinsetterAsyncInterceptor.class);
+        bind(PinsetterAsyncFilter.class);
         bind(VersionResponseFilter.class);
         bind(LinkHeaderResponseFilter.class);
         bind(DynamicJsonFilter.class);
