+++ server/src/main/java/org/candlepin/resteasy/filter/VerifyAuthorizationFilter.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.resteasy.interceptor;
+package org.candlepin.resteasy.filter;
 
 import org.candlepin.auth.Access;
 import org.candlepin.auth.Principal;
