+++ server/src/test/java/org/candlepin/resteasy/filter/VerifyAuthorizationFilterTest.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.resteasy.interceptor;
+package org.candlepin.resteasy.filter;
 
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.*;
